fix: changed type price to correct formatting

+++ document/internal/dto/consultancy-dto.go
@@ -1,5 +1,5 @@
 package dto
 
 type consultancyDTO struct {
-	Price      float64
+	Price      numeric
 }
